add the managed cluster info controller
