+++ agent/pkg/status/controller/managedclusters/k8s_cluster_test.go
@@ -1,4 +1,4 @@
-package transfer
+package managedclusters
 
 import (
 	"testing"
