+++ pkg/transport/requester/inventory_client_test.go
@@ -1,4 +1,4 @@
-package client
+package requester
 
 import (
 	"context"
